fluid clean layer
