+++ python/paddle/fluid/layers/rnn.py
@@ -36,501 +36,13 @@ from ..data_feeder import check_variable_and_dtype, check_type, check_dtype
 from collections.abc import Sequence
 
 __all__ = [
-    'RNNCell',
-    'GRUCell',
-    'LSTMCell',
-    'rnn',
-    'birnn',
     'dynamic_decode',
     'dynamic_lstm',
     'dynamic_lstmp',
     'dynamic_gru',
     'gru_unit',
-    'lstm_unit',
     'lstm',
 ]
-
-
-class RNNCell:
-    """
-        :api_attr: Static Graph
-
-    RNNCell is the base class for abstraction representing the calculations
-    mapping the input and state to the output and new state. It is suitable to
-    and mostly used in RNN.
-    """
-
-    def call(self, inputs, states, **kwargs):
-        r"""
-        Every cell must implement this method to do the calculations mapping the
-        inputs and states to the output and new states.
-
-        To be more flexible, both inputs and states can be a tensor variable or
-        a nested structure (list|tuple|namedtuple|dict) of tensor variable, that
-        is, a (possibly nested structure of) tensor variable[s].
-
-        Parameters:
-            inputs: A (possibly nested structure of) tensor variable[s].
-            states: A (possibly nested structure of) tensor variable[s].
-            **kwargs: Additional keyword arguments, provided by the caller.
-
-        Returns:
-            tuple: outputs and new_states pair. outputs and new_states both \
-                can be nested structure of tensor variables. new_states must \
-                have the same structure with states.
-
-        """
-        raise NotImplementedError("RNNCell must implent the call function.")
-
-    def __call__(self, inputs, states, **kwargs):
-        return self.call(inputs, states, **kwargs)
-
-    def get_initial_states(
-        self,
-        batch_ref,
-        shape=None,
-        dtype='float32',
-        init_value=0,
-        batch_dim_idx=0,
-    ):
-        r"""
-        Generate initialized states according to provided shape, data type and
-        value.
-
-        Parameters:
-            batch_ref: A (possibly nested structure of) tensor variable[s].
-                The first dimension of the tensor will be used as batch size to
-                initialize states.
-            shape: A (possibly nested structure of) shape[s], where a shape is
-                represented as a list/tuple of integer). -1(for batch size) will
-                beautomatically inserted if shape is not started with it. If None,
-                property `state_shape` will be used. The default value is None.
-            dtype: A (possibly nested structure of) data type[s]. The structure
-                must be same as that of `shape`, except when all tensors' in states
-                has the same data type, a single data type can be used. If
-                property `cell.state_shape` is not available, float32 will be used
-                as the data type. The default value is float32.
-            init_value: A float value used to initialize states.
-            batch_dim_idx: An integer indicating which dimension of the tensor in
-                inputs represents batch size.  The default value is 0.
-
-        Returns:
-            Variable: tensor variable[s] packed in the same structure provided \
-                by shape, representing the initialized states.
-        """
-        check_variable_and_dtype(
-            batch_ref,
-            'batch_ref',
-            ['float32', 'float64', 'int32', 'int64'],
-            'RNNCell',
-        )
-        check_type(shape, 'shape', (list, tuple, type(None), int), 'RNNCell')
-        if isinstance(shape, (list, tuple)):
-            shapes = map_structure(lambda x: x, shape)
-            if isinstance(shape, list):
-                for i, _shape in enumerate(shapes):
-                    check_type(_shape, 'shapes[' + str(i) + ']', int, 'RNNCell')
-            else:
-                check_type(shapes, 'shapes', int, 'RNNCell')
-        check_dtype(dtype, 'dtype', ['float32', 'float64'], 'RNNCell')
-
-        # TODO: use inputs and batch_size
-        batch_ref = flatten(batch_ref)[0]
-
-        def _is_shape_sequence(seq):
-            """For shape, list/tuple of integer is the finest-grained objection"""
-            if isinstance(seq, list) or isinstance(seq, tuple):
-                if reduce(
-                    lambda flag, x: isinstance(x, int) and flag, seq, True
-                ):
-                    return False
-            # TODO: Add check for the illegal
-            if isinstance(seq, dict):
-                return True
-            return isinstance(seq, Sequence) and not isinstance(seq, str)
-
-        class Shape:
-            def __init__(self, shape):
-                self.shape = shape if shape[0] == -1 else ([-1] + list(shape))
-
-        # nested structure of shapes
-        states_shapes = self.state_shape if shape is None else shape
-        is_sequence_ori = utils.is_sequence
-        utils.is_sequence = _is_shape_sequence
-        states_shapes = map_structure(lambda shape: Shape(shape), states_shapes)
-        utils.is_sequence = is_sequence_ori
-
-        # nested structure of dtypes
-        try:
-            states_dtypes = self.state_dtype if dtype is None else dtype
-        except NotImplementedError:  # use fp32 as default
-            states_dtypes = "float32"
-        if len(flatten(states_dtypes)) == 1:
-            dtype = flatten(states_dtypes)[0]
-            states_dtypes = map_structure(lambda shape: dtype, states_shapes)
-
-        init_states = map_structure(
-            lambda shape, dtype: tensor.fill_constant_batch_size_like(
-                input=batch_ref,
-                shape=shape.shape,
-                dtype=dtype,
-                value=init_value,
-                input_dim_idx=batch_dim_idx,
-            ),
-            states_shapes,
-            states_dtypes,
-        )
-        return init_states
-
-    @property
-    def state_shape(self):
-        """
-        Abstract method (property).
-        Used to initialize states.
-        A (possibly nested structure of) shape[s], where a shape is represented
-        as a list/tuple of integers (-1 for batch size would be automatically
-        inserted into a shape if shape is not started with it).
-        Not necessary to be implemented if states are not initialized by
-        `get_initial_states` or the `shape` argument is provided when using
-        `get_initial_states`.
-        """
-        raise NotImplementedError(
-            "Please add implementaion for `state_shape` in the used cell."
-        )
-
-    @property
-    def state_dtype(self):
-        """
-        Abstract method (property).
-        Used to initialize states.
-        A (possibly nested structure of) data types[s]. The structure must be
-        same as that of `shape`, except when all tensors' in states has the same
-        data type, a single data type can be used.
-        Not necessary to be implemented if states are not initialized
-        by `get_initial_states` or the `dtype` argument is provided when using
-        `get_initial_states`.
-        """
-        raise NotImplementedError(
-            "Please add implementaion for `state_dtype` in the used cell."
-        )
-
-
-class GRUCell(RNNCell):
-    r"""
-        :api_attr: Static Graph
-
-    Gated Recurrent Unit cell. It is a wrapper for
-    `fluid.contrib.layers.rnn_impl.BasicGRUUnit` to make it adapt to RNNCell.
-
-    The formula used is as follow:
-
-    .. math::
-
-        u_t & = act_g(W_{ux}x_{t} + W_{uh}h_{t-1} + b_u)
-
-        r_t & = act_g(W_{rx}x_{t} + W_{rh}h_{t-1} + b_r)
-
-        \\tilde{h_t} & = act_c(W_{cx}x_{t} + W_{ch}(r_t \odot h_{t-1}) + b_c)
-
-        h_t & = u_t \odot h_{t-1} + (1-u_t) \odot \\tilde{h_t}
-
-    For more details, please refer to  `Learning Phrase Representations using
-    RNN Encoder Decoder for Statistical Machine Translation <https://arxiv.org/pdf/1406.1078.pdf>`_
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle.fluid.layers as layers
-            cell = layers.GRUCell(hidden_size=256)
-    """
-
-    def __init__(
-        self,
-        hidden_size,
-        param_attr=None,
-        bias_attr=None,
-        gate_activation=None,
-        activation=None,
-        dtype="float32",
-        name="GRUCell",
-    ):
-        """
-        Constructor of GRUCell.
-
-        Parameters:
-            hidden_size (int): The hidden size in the GRU cell.
-            param_attr(ParamAttr, optional): The parameter attribute for the learnable
-                weight matrix. Default: None.
-            bias_attr (ParamAttr, optional): The parameter attribute for the bias
-                of GRU. Default: None.
-            gate_activation (function, optional): The activation function for :math:`act_g`.
-                Default: `fluid.layers.sigmoid`.
-            activation (function, optional): The activation function for :math:`act_c`.
-                Default: `fluid.layers.tanh`.
-            dtype(string, optional): The data type used in this cell. Default float32.
-            name(string, optional) : The name scope used to identify parameters and biases.
-        """
-        check_type(hidden_size, 'hidden_size', (int), 'GRUCell')
-        check_dtype(dtype, 'dtype', ['float32', 'float64'], 'GRUCell')
-        self.hidden_size = hidden_size
-        from .. import contrib  # TODO: resolve recurrent import
-
-        self.gru_unit = contrib.layers.rnn_impl.BasicGRUUnit(
-            name,
-            hidden_size,
-            param_attr,
-            bias_attr,
-            gate_activation,
-            activation,
-            dtype,
-        )
-
-    def call(self, inputs, states):
-        r"""
-        Perform calculations of GRU.
-
-        Parameters:
-            inputs(Variable): A tensor with shape `[batch_size, input_size]`,
-                corresponding to :math:`x_t` in the formula. The data type
-                should be float32 or float64.
-            states(Variable): A tensor with shape `[batch_size, hidden_size]`.
-                corresponding to :math:`h_{t-1}` in the formula. The data type
-                should be float32 or float64.
-
-        Returns:
-            tuple: A tuple( :code:`(outputs, new_states)` ), where `outputs` and \
-                `new_states` is the same tensor shaped `[batch_size, hidden_size]`, \
-                corresponding to :math:`h_t` in the formula. The data type of the \
-                tensor is same as that of `states`.
-        """
-
-        check_variable_and_dtype(
-            inputs, 'inputs', ['float32', 'float64'], 'GRUCell'
-        )
-        check_variable_and_dtype(
-            states, 'states', ['float32', 'float64'], 'GRUCell'
-        )
-        new_hidden = self.gru_unit(inputs, states)
-        return new_hidden, new_hidden
-
-    @property
-    def state_shape(self):
-        """
-        The `state_shape` of GRUCell is a shape `[hidden_size]` (-1 for batch
-        size would be automatically inserted into shape). The shape corresponds
-        to :math:`h_{t-1}`.
-        """
-        return [self.hidden_size]
-
-
-class LSTMCell(RNNCell):
-    r"""
-        :api_attr: Static Graph
-
-    Long-Short Term Memory cell. It is a wrapper for
-    `fluid.contrib.layers.rnn_impl.BasicLSTMUnit` to make it adapt to RNNCell.
-
-    The formula used is as follow:
-
-    .. math::
-
-        i_{t} & = act_g(W_{x_{i}}x_{t} + W_{h_{i}}h_{t-1} + b_{i})
-
-        f_{t} & = act_g(W_{x_{f}}x_{t} + W_{h_{f}}h_{t-1} + b_{f} + forget\\_bias)
-
-        c_{t} & = f_{t}c_{t-1} + i_{t} act_c (W_{x_{c}}x_{t} + W_{h_{c}}h_{t-1} + b_{c})
-
-        o_{t} & = act_g(W_{x_{o}}x_{t} + W_{h_{o}}h_{t-1} + b_{o})
-
-        h_{t} & = o_{t} act_c (c_{t})
-
-    For more details, please refer to `RECURRENT NEURAL NETWORK REGULARIZATION <http://arxiv.org/abs/1409.2329>`_
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle.fluid.layers as layers
-            cell = layers.LSTMCell(hidden_size=256)
-    """
-
-    def __init__(
-        self,
-        hidden_size,
-        param_attr=None,
-        bias_attr=None,
-        gate_activation=None,
-        activation=None,
-        forget_bias=1.0,
-        dtype="float32",
-        name="LSTMCell",
-    ):
-        """
-        Constructor of LSTMCell.
-
-        Parameters:
-            hidden_size (int): The hidden size in the LSTM cell.
-            param_attr(ParamAttr, optional): The parameter attribute for the learnable
-                weight matrix. Default: None.
-            bias_attr (ParamAttr, optional): The parameter attribute for the bias
-                of LSTM. Default: None.
-            gate_activation (function, optional): The activation function for :math:`act_g`.
-                Default: 'fluid.layers.sigmoid'.
-            activation (function, optional): The activation function for :math:`act_h`.
-                Default: 'fluid.layers.tanh'.
-            forget_bias(float, optional): forget bias used when computing forget gate.
-                Default 1.0
-            dtype(string, optional): The data type used in this cell. Default float32.
-            name(string, optional) : The name scope used to identify parameters and biases.
-        """
-
-        check_type(hidden_size, 'hidden_size', (int), 'LSTMCell')
-        check_dtype(dtype, 'dtype', ['float32', 'float64'], 'LSTMCell')
-        self.hidden_size = hidden_size
-        from .. import contrib  # TODO: resolve recurrent import
-
-        self.lstm_unit = contrib.layers.rnn_impl.BasicLSTMUnit(
-            name,
-            hidden_size,
-            param_attr,
-            bias_attr,
-            gate_activation,
-            activation,
-            forget_bias,
-            dtype,
-        )
-
-    def call(self, inputs, states):
-        r"""
-        Perform calculations of LSTM.
-
-        Parameters:
-            inputs(Variable): A tensor with shape `[batch_size, input_size]`,
-                corresponding to :math:`x_t` in the formula. The data type
-                should be float32 or float64.
-            states(Variable): A list of containing two tensors, each shaped
-                `[batch_size, hidden_size]`, corresponding to :math:`h_{t-1}, c_{t-1}`
-                in the formula. The data type should be float32 or float64.
-
-        Returns:
-            tuple: A tuple( :code:`(outputs, new_states)` ), where `outputs` is \
-                a tensor with shape `[batch_size, hidden_size]`, corresponding \
-                to :math:`h_{t}` in the formula; `new_states` is a list containing \
-                two tenser variables shaped `[batch_size, hidden_size]`, corresponding \
-                to :math:`h_{t}, c_{t}` in the formula. The data type of these \
-                tensors all is same as that of `states`.
-        """
-
-        check_variable_and_dtype(
-            inputs, 'inputs', ['float32', 'float64'], 'LSTMCell'
-        )
-        check_type(states, 'states', list, 'LSTMCell')
-        if isinstance(states, list):
-            for i, state in enumerate(states):
-                check_variable_and_dtype(
-                    state,
-                    'state[' + str(i) + ']',
-                    ['float32', 'float64'],
-                    'LSTMCell',
-                )
-
-        pre_hidden, pre_cell = states
-        new_hidden, new_cell = self.lstm_unit(inputs, pre_hidden, pre_cell)
-        return new_hidden, [new_hidden, new_cell]
-
-    @property
-    def state_shape(self):
-        """
-        The `state_shape` of LSTMCell is a list with two shapes: `[[hidden_size], [hidden_size]]`
-        (-1 for batch size would be automatically inserted into shape). These two
-        shapes correspond to :math:`h_{t-1}` and :math:`c_{t-1}` separately.
-        """
-        return [[self.hidden_size], [self.hidden_size]]
-
-
-def rnn(
-    cell,
-    inputs,
-    initial_states=None,
-    sequence_length=None,
-    time_major=False,
-    is_reverse=False,
-    **kwargs
-):
-    """
-    rnn creates a recurrent neural network specified by RNNCell `cell`,
-    which performs :code:`cell.call()` (for dygraph mode :code:`cell.forward`)
-    repeatedly until reaches to the maximum length of `inputs`.
-
-    Arguments:
-        cell(RNNCellBase): An instance of `RNNCellBase`.
-        inputs(Tensor): the input sequences.
-            If time_major is True, the shape is
-            `[time_steps, batch_size, input_size]`
-            else the shape is `[batch_size, time_steps, input_size]`.
-        initial_states(Tensor|tuple|list, optional): the initial state of the
-            rnn cell. Tensor or a possibly nested structure of tensors. If not
-            provided, `cell.get_initial_states` would be called to produce
-            the initial state. Defaults to None.
-        sequence_length (Tensor, optional): shape `[batch_size]`, dtype: int64
-            or int32. The valid lengths of input sequences. Defaults to None.
-            If `sequence_length` is not None, the inputs are treated as
-            padded sequences. In each input sequence, elements whose time step
-            index are not less than the valid length are treated as paddings.
-        time_major (bool): Whether the first dimension of the input means the
-            time steps. Defaults to False.
-        is_reverse (bool, optional): Indicate whether to calculate in the reverse
-            order of input sequences. Defaults to False.
-        **kwargs: Additional keyword arguments to pass to `forward` of the cell.
-
-    Returns:
-        (outputs, final_states)
-        outputs (Tensor|list|tuple): the output sequence. Tensor or nested
-            structure of Tensors.
-            If `time_major` is True, the shape of each tensor in outpus is
-            `[time_steps, batch_size, hidden_size]`, else
-            `[batch_size, time_steps, hidden_size]`.
-        final_states (Tensor|list|tuple): final states. A (possibly nested structure of)
-            tensor[s], representing the final state for RNN. It has the same
-            structure of intial state. Each tensor in final states has the same
-            shape and dtype as the corresponding tensor in initial states.
-
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle
-            paddle.disable_static()
-
-            cell = paddle.nn.SimpleRNNCell(16, 32)
-
-            inputs = paddle.rand((4, 23, 16))
-            prev_h = paddle.randn((4, 32))
-            outputs, final_states = paddle.fluid.layers.rnn(cell, inputs, prev_h)
-
-    """
-    if _non_static_mode():
-        return _rnn_dynamic_graph(
-            cell,
-            inputs,
-            initial_states,
-            sequence_length,
-            time_major,
-            is_reverse,
-            **kwargs
-        )
-    else:
-        return _rnn_static_graph(
-            cell,
-            inputs,
-            initial_states,
-            sequence_length,
-            time_major,
-            is_reverse,
-            **kwargs
-        )
 
 
 class ArrayWrapper:
@@ -543,273 +55,6 @@ class ArrayWrapper:
 
     def __getitem__(self, item):
         return self.array.__getitem__(item)
-
-
-def _maybe_copy(state, new_state, step_mask):
-    """update rnn state or just pass the old state through"""
-    new_state = nn.elementwise_mul(
-        new_state, step_mask, axis=0
-    ) + nn.elementwise_mul(state, (1 - step_mask), axis=0)
-    return new_state
-
-
-def _transpose_batch_time(x):
-    perm = [1, 0] + list(range(2, len(x.shape)))
-    return paddle.transpose(x, perm)
-
-
-def _rnn_dynamic_graph(
-    cell,
-    inputs,
-    initial_states=None,
-    sequence_length=None,
-    time_major=False,
-    is_reverse=False,
-    **kwargs
-):
-    time_step_index = 0 if time_major else 1
-    flat_inputs = flatten(inputs)
-    time_steps = flat_inputs[0].shape[time_step_index]
-
-    if initial_states is None:
-        initial_states = cell.get_initial_states(
-            batch_ref=inputs, batch_dim_idx=1 if time_major else 0
-        )
-
-    if not time_major:
-        inputs = map_structure(_transpose_batch_time, inputs)
-
-    if sequence_length is not None:
-        mask = sequence_lod.sequence_mask(
-            sequence_length, maxlen=time_steps, dtype=inputs.dtype
-        )
-        mask = paddle.transpose(mask, [1, 0])
-
-    if is_reverse:
-        inputs = map_structure(lambda x: paddle.reverse(x, axis=[0]), inputs)
-        mask = (
-            paddle.reverse(mask, axis=[0])
-            if sequence_length is not None
-            else None
-        )
-
-    states = initial_states
-    outputs = []
-    for i in range(time_steps):
-        step_inputs = map_structure(lambda x: x[i], inputs)
-        step_outputs, new_states = cell(step_inputs, states, **kwargs)
-        if sequence_length is not None:
-            new_states = map_structure(
-                partial(_maybe_copy, step_mask=mask[i]), states, new_states
-            )
-        states = new_states
-        outputs = (
-            map_structure(lambda x: ArrayWrapper(x), step_outputs)
-            if i == 0
-            else map_structure(
-                lambda x, x_array: x_array.append(x), step_outputs, outputs
-            )
-        )
-
-    final_outputs = map_structure(
-        lambda x: paddle.stack(x.array, axis=time_step_index), outputs
-    )
-
-    if is_reverse:
-        final_outputs = map_structure(
-            lambda x: paddle.reverse(x, axis=time_step_index), final_outputs
-        )
-
-    final_states = new_states
-    return final_outputs, final_states
-
-
-def _rnn_static_graph(
-    cell,
-    inputs,
-    initial_states=None,
-    sequence_length=None,
-    time_major=False,
-    is_reverse=False,
-    **kwargs
-):
-    check_type(inputs, 'inputs', (Variable, list, tuple), 'rnn')
-    if isinstance(inputs, (list, tuple)):
-        for i, input_x in enumerate(inputs):
-            check_variable_and_dtype(
-                input_x, 'inputs[' + str(i) + ']', ['float32', 'float64'], 'rnn'
-            )
-    check_type(
-        initial_states,
-        'initial_states',
-        (Variable, list, tuple, type(None)),
-        'rnn',
-    )
-
-    check_type(
-        sequence_length, 'sequence_length', (Variable, type(None)), 'rnn'
-    )
-
-    def _switch_grad(x, stop=False):
-        x.stop_gradient = stop
-        return x
-
-    if initial_states is None:
-        initial_states = cell.get_initial_states(
-            batch_ref=inputs, batch_dim_idx=1 if time_major else 0
-        )
-    initial_states = map_structure(_switch_grad, initial_states)
-
-    if not time_major:
-        inputs = map_structure(_transpose_batch_time, inputs)
-
-    if sequence_length:
-        max_seq_len = paddle.shape(flatten(inputs)[0])[0]
-        mask = sequence_lod.sequence_mask(
-            sequence_length,
-            maxlen=max_seq_len,
-            dtype=flatten(initial_states)[0].dtype,
-        )
-        mask = paddle.transpose(mask, [1, 0])
-    if is_reverse:
-        inputs = map_structure(lambda x: paddle.reverse(x, axis=[0]), inputs)
-        mask = paddle.reverse(mask, axis=[0]) if sequence_length else None
-
-    # StaticRNN
-    rnn = control_flow.StaticRNN()
-    with rnn.step():
-        inputs = map_structure(rnn.step_input, inputs)
-        states = map_structure(rnn.memory, initial_states)
-        copy_states = map_structure(lambda x: x, states)
-        outputs, new_states = cell(inputs, copy_states, **kwargs)
-        assert_same_structure(states, new_states)
-        if sequence_length:
-            step_mask = rnn.step_input(mask)
-            new_states = map_structure(
-                partial(_maybe_copy, step_mask=step_mask), states, new_states
-            )
-
-        map_structure(rnn.update_memory, states, new_states)
-        flat_outputs = flatten(outputs)
-        map_structure(rnn.step_output, outputs)
-        map_structure(rnn.step_output, new_states)
-
-    rnn_out = rnn()
-    final_outputs = rnn_out[: len(flat_outputs)]
-    final_outputs = pack_sequence_as(outputs, final_outputs)
-    final_states = map_structure(lambda x: x[-1], rnn_out[len(flat_outputs) :])
-    final_states = pack_sequence_as(new_states, final_states)
-
-    if is_reverse:
-        final_outputs = map_structure(
-            lambda x: paddle.reverse(x, axis=[0]), final_outputs
-        )
-
-    if not time_major:
-        final_outputs = map_structure(_transpose_batch_time, final_outputs)
-
-    return (final_outputs, final_states)
-
-
-def birnn(
-    cell_fw,
-    cell_bw,
-    inputs,
-    initial_states=None,
-    sequence_length=None,
-    time_major=False,
-    **kwargs
-):
-    """
-    birnn creates a bidirectional recurrent neural network specified by
-    RNNCell `cell_fw` and `cell_bw`, which performs :code:`cell.call()`
-    (for dygraph mode :code:`cell.forward`) repeatedly until reaches to
-    the maximum length of `inputs` and then concat the outputs for both RNNs
-    along the last axis.
-
-    Arguments:
-        cell_fw(RNNCellBase): An instance of `RNNCellBase`.
-        cell_bw(RNNCellBase): An instance of `RNNCellBase`.
-        inputs(Tensor): the input sequences.
-            If time_major is True, the shape is
-            `[time_steps, batch_size, input_size]`
-            else the shape is `[batch_size, time_steps, input_size]`.
-        initial_states(tuple, optional): A tuple of initial states of
-            `cell_fw` and `cell_bw`.
-            If not provided, `cell.get_initial_states` would be called to
-            produce initial state for each cell. Defaults to None.
-        sequence_length (Tensor, optional): shape `[batch_size]`, dtype: int64
-            or int32. The valid lengths of input sequences. Defaults to None.
-            If `sequence_length` is not None, the inputs are treated as
-            padded sequences. In each input sequence, elements whose time step
-            index are not less than the valid length are treated as paddings.
-        time_major (bool): Whether the first dimension of the input means the
-            time steps. Defaults to False.
-        **kwargs: Additional keyword arguments to pass to `forward` of each cell.
-
-    Returns:
-        (outputs, final_states)
-        outputs (Tensor): the outputs of the bidirectional RNN. It is the
-            concatenation of the outputs from the forward RNN and backward
-            RNN along the last axis.
-            If time major is True, the shape is `[time_steps, batch_size, size]`,
-            else the shape is `[batch_size, time_steps, size]`, where size is
-            `cell_fw.hidden_size + cell_bw.hidden_size`.
-        final_states (tuple): A tuple of the final states of the forward
-            cell and backward cell.
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle
-            paddle.disable_static()
-
-            cell_fw = paddle.nn.LSTMCell(16, 32)
-            cell_bw = paddle.nn.LSTMCell(16, 32)
-
-            inputs = paddle.rand((4, 23, 16))
-            hf, cf = paddle.rand((4, 32)), paddle.rand((4, 32))
-            hb, cb = paddle.rand((4, 32)), paddle.rand((4, 32))
-            initial_states = ((hf, cf), (hb, cb))
-            outputs, final_states = paddle.fluid.layers.birnn(
-                cell_fw, cell_bw, inputs, initial_states)
-
-    """
-    if initial_states is None:
-        states_fw = cell_fw.get_initial_states(
-            batch_ref=inputs, batch_dim_idx=1 if time_major else 0
-        )
-        states_bw = cell_fw.get_initial_states(
-            batch_ref=inputs, batch_dim_idx=1 if time_major else 0
-        )
-    else:
-        states_fw, states_bw = initial_states
-    outputs_fw, states_fw = rnn(
-        cell_fw,
-        inputs,
-        states_fw,
-        sequence_length,
-        time_major=time_major,
-        **kwargs
-    )
-
-    outputs_bw, states_bw = rnn(
-        cell_bw,
-        inputs,
-        states_bw,
-        sequence_length,
-        time_major=time_major,
-        is_reverse=True,
-        **kwargs
-    )
-
-    outputs = map_structure(
-        lambda x, y: tensor.concat([x, y], -1), outputs_fw, outputs_bw
-    )
-
-    final_states = (states_fw, states_bw)
-    return outputs, final_states
 
 
 def _dynamic_decode_imperative(
@@ -2170,151 +1415,3 @@ def gru_unit(
     )
 
     return updated_hidden, reset_hidden_pre, gate
-
-
-def lstm_unit(
-    x_t,
-    hidden_t_prev,
-    cell_t_prev,
-    forget_bias=0.0,
-    param_attr=None,
-    bias_attr=None,
-    name=None,
-):
-    r"""
-	:api_attr: Static Graph
-
-    Long-Short Term Memory (LSTM) RNN cell. This operator performs LSTM calculations for
-    one time step, whose implementation is based on calculations described in `RECURRENT
-    NEURAL NETWORK REGULARIZATION <http://arxiv.org/abs/1409.2329>`_  .
-
-    We add forget_bias to the biases of the forget gate in order to
-    reduce the scale of forgetting. The formula is as follows:
-
-    .. math::
-
-        i_{t} & = \sigma(W_{x_{i}}x_{t} + W_{h_{i}}h_{t-1} + b_{i})
-
-        f_{t} & = \sigma(W_{x_{f}}x_{t} + W_{h_{f}}h_{t-1} + b_{f} + forget\\_bias)
-
-        c_{t} & = f_{t}c_{t-1} + i_{t} tanh (W_{x_{c}}x_{t} + W_{h_{c}}h_{t-1} + b_{c})
-
-        o_{t} & = \sigma(W_{x_{o}}x_{t} + W_{h_{o}}h_{t-1} + b_{o})
-
-        h_{t} & = o_{t} tanh (c_{t})
-
-    :math:`x_{t}` stands for ``x_t`` , corresponding to the input of current time step;
-    :math:`h_{t-1}` and :math:`c_{t-1}` correspond to ``hidden_t_prev`` and ``cell_t_prev`` ,
-    representing the output of from previous time step.
-    :math:`i_{t}, f_{t}, c_{t}, o_{t}, h_{t}` are input gate, forget gate, cell, output gate
-    and hidden calculation.
-
-    Args:
-        x_t(Variable): A 2D Tensor representing the input of current time step.
-            Its shape should be :math:`[N, M]` , where :math:`N` stands for batch
-            size, :math:`M` for the feature size of input. The data type should
-            be float32 or float64.
-        hidden_t_prev(Variable): A 2D Tensor representing the hidden value from
-            previous step. Its shape should be :math:`[N, D]` , where :math:`N`
-            stands for batch size, :math:`D` for the hidden size. The data type
-            should be same as ``x_t`` .
-        cell_t_prev(Variable): A 2D Tensor representing the cell value from
-            previous step. It has the same shape and data type with ``hidden_t_prev`` .
-        forget_bias (float, optional): :math:`forget\\_bias` added to the biases
-            of the forget gate. Default 0.
-        param_attr(ParamAttr, optional):  To specify the weight parameter property.
-            Default: None, which means the default weight parameter property is used.
-            See usage for details in :ref:`api_fluid_ParamAttr` .
-        bias_attr (ParamAttr, optional): To specify the bias parameter property.
-            Default: None, which means the default bias parameter property is used.
-            See usage for details in :ref:`api_fluid_ParamAttr` .
-        name(str, optional): For detailed information, please refer
-            to :ref:`api_guide_Name`. Usually name is no need to set and
-            None by default.
-
-    Returns:
-        tuple: The tuple contains two Tensor variables with the same shape and \
-            data type with ``hidden_t_prev`` , representing the hidden value and \
-            cell value which correspond to :math:`h_{t}` and :math:`c_{t}` in \
-            the formula.
-
-    Raises:
-        ValueError: Rank of x_t must be 2.
-        ValueError: Rank of hidden_t_prev must be 2.
-        ValueError: Rank of cell_t_prev must be 2.
-        ValueError: The 1st dimensions of x_t, hidden_t_prev and cell_t_prev must be the same.
-        ValueError: The 2nd dimensions of hidden_t_prev and cell_t_prev must be the same.
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle.fluid as fluid
-
-            dict_dim, emb_dim, hidden_dim = 128, 64, 512
-            data = fluid.data(name='step_data', shape=[None], dtype='int64')
-            x = fluid.embedding(input=data, size=[dict_dim, emb_dim])
-            pre_hidden = fluid.data(
-                name='pre_hidden', shape=[None, hidden_dim], dtype='float32')
-            pre_cell = fluid.data(
-                name='pre_cell', shape=[None, hidden_dim], dtype='float32')
-            hidden = fluid.layers.lstm_unit(
-                x_t=x,
-                hidden_t_prev=pre_hidden,
-                cell_t_prev=pre_cell)
-    """
-    helper = LayerHelper('lstm_unit', **locals())
-    check_variable_and_dtype(x_t, 'x_t', ['float32', 'float64'], 'lstm_unit')
-    check_variable_and_dtype(
-        hidden_t_prev, 'hidden_t_prev', ['float32', 'float64'], 'lstm_unit'
-    )
-    check_variable_and_dtype(
-        cell_t_prev, 'cell_t_prev', ['float32', 'float64'], 'lstm_unit'
-    )
-    if len(x_t.shape) != 2:
-        raise ValueError("Rank of x_t must be 2.")
-
-    if len(hidden_t_prev.shape) != 2:
-        raise ValueError("Rank of hidden_t_prev must be 2.")
-
-    if len(cell_t_prev.shape) != 2:
-        raise ValueError("Rank of cell_t_prev must be 2.")
-
-    if (
-        x_t.shape[0] != hidden_t_prev.shape[0]
-        or x_t.shape[0] != cell_t_prev.shape[0]
-    ):
-        raise ValueError(
-            "The 1st dimensions of x_t, hidden_t_prev and "
-            "cell_t_prev must be the same."
-        )
-
-    if hidden_t_prev.shape[1] != cell_t_prev.shape[1]:
-        raise ValueError(
-            "The 2nd dimensions of hidden_t_prev and "
-            "cell_t_prev must be the same."
-        )
-
-    if bias_attr is None:
-        bias_attr = ParamAttr()
-
-    size = cell_t_prev.shape[1]
-    concat_out = nn.concat(input=[x_t, hidden_t_prev], axis=1)
-    fc_out = nn.fc(
-        input=concat_out,
-        size=4 * size,
-        param_attr=param_attr,
-        bias_attr=bias_attr,
-    )
-    dtype = x_t.dtype
-    c = helper.create_variable_for_type_inference(dtype)
-    h = helper.create_variable_for_type_inference(dtype)
-
-    helper.append_op(
-        type='lstm_unit',
-        inputs={"X": fc_out, "C_prev": cell_t_prev},
-        outputs={"C": c, "H": h},
-        attrs={"forget_bias": forget_bias},
-    )
-
-    return h, c
