+++ python/paddle/fluid/tests/unittests/test_lstm_unit_op.py
@@ -17,10 +17,6 @@ import unittest
 import numpy as np
 from op_test import OpTest
 
-from paddle import fluid
-from paddle.fluid.framework import Program, program_guard
-from paddle.fluid.layers import lstm_unit
-
 
 def sigmoid_np(x):
     return 1.0 / (1.0 + np.exp(-x))
@@ -28,79 +24,6 @@ def sigmoid_np(x):
 
 def tanh_np(x):
     return 2 * sigmoid_np(2.0 * x) - 1.0
-
-
-class LstmUnitTestError(unittest.TestCase):
-    def test_errors(self):
-        with program_guard(Program(), Program()):
-            batch_size, dict_dim, emb_dim, hidden_dim = 32, 128, 64, 512
-            data = fluid.data(
-                name='step_data', shape=[batch_size], dtype='int64'
-            )
-            inputs = fluid.embedding(input=data, size=[dict_dim, emb_dim])
-            pre_hidden = fluid.data(
-                name='pre_hidden',
-                shape=[batch_size, hidden_dim],
-                dtype='float32',
-            )
-            pre_cell = fluid.data(
-                name='pre_cell', shape=[batch_size, hidden_dim], dtype='float32'
-            )
-
-            np_input = np.random.uniform(
-                -0.1, 0.1, (batch_size, emb_dim)
-            ).astype('float64')
-            np_pre_hidden = np.random.uniform(
-                -0.1, 0.1, (batch_size, hidden_dim)
-            ).astype('float64')
-            np_pre_cell = np.random.uniform(
-                -0.1, 0.1, (batch_size, hidden_dim)
-            ).astype('float64')
-
-            def test_input_Variable():
-                lstm_unit(np_input, pre_hidden, pre_cell)
-
-            self.assertRaises(TypeError, test_input_Variable)
-
-            def test_pre_hidden_Variable():
-                lstm_unit(inputs, np_pre_hidden, pre_cell)
-
-            self.assertRaises(TypeError, test_pre_hidden_Variable)
-
-            def test_pre_cell_Variable():
-                lstm_unit(inputs, pre_hidden, np_pre_cell)
-
-            self.assertRaises(TypeError, test_pre_cell_Variable)
-
-            def test_input_type():
-                error_input = fluid.data(
-                    name='error_input',
-                    shape=[batch_size, emb_dim],
-                    dtype='int32',
-                )
-                lstm_unit(error_input, pre_hidden, pre_cell)
-
-            self.assertRaises(TypeError, test_input_type)
-
-            def test_pre_hidden_type():
-                error_pre_hidden = fluid.data(
-                    name='error_pre_hidden',
-                    shape=[batch_size, hidden_dim],
-                    dtype='int32',
-                )
-                lstm_unit(inputs, error_pre_hidden, pre_cell)
-
-            self.assertRaises(TypeError, test_pre_hidden_type)
-
-            def test_pre_cell_type():
-                error_pre_cell = fluid.data(
-                    name='error_pre_cell',
-                    shape=[batch_size, hidden_dim],
-                    dtype='int32',
-                )
-                lstm_unit(inputs, pre_hidden, error_pre_cell)
-
-            self.assertRaises(TypeError, test_pre_cell_type)
 
 
 class LstmUnitTest(OpTest):
