+++ python/paddle/fluid/tests/unittests/test_rnn_decode_api.py
@@ -19,268 +19,15 @@ import numpy as np
 
 import paddle
 import paddle.fluid as fluid
-import paddle.fluid.core as core
 import paddle.fluid.layers as layers
 import paddle.nn as nn
 from paddle import Model, set_device
 from paddle.fluid.dygraph import Layer
-from paddle.fluid.executor import Executor
 from paddle.fluid.framework import _test_eager_guard
 from paddle.nn import BeamSearchDecoder, dynamic_decode
 from paddle.static import InputSpec as Input
 
 paddle.enable_static()
-
-
-class EncoderCell(layers.RNNCell):
-    def __init__(self, num_layers, hidden_size, dropout_prob=0.0):
-        self.num_layers = num_layers
-        self.hidden_size = hidden_size
-        self.dropout_prob = dropout_prob
-        self.lstm_cells = [
-            layers.LSTMCell(hidden_size) for i in range(num_layers)
-        ]
-
-    def call(self, step_input, states):
-        new_states = []
-        for i in range(self.num_layers):
-            out, new_state = self.lstm_cells[i](step_input, states[i])
-            step_input = (
-                layers.dropout(out, self.dropout_prob)
-                if self.dropout_prob > 0
-                else out
-            )
-            new_states.append(new_state)
-        return step_input, new_states
-
-    @property
-    def state_shape(self):
-        return [cell.state_shape for cell in self.lstm_cells]
-
-
-class DecoderCell(layers.RNNCell):
-    def __init__(self, num_layers, hidden_size, dropout_prob=0.0):
-        self.num_layers = num_layers
-        self.hidden_size = hidden_size
-        self.dropout_prob = dropout_prob
-        self.lstm_cells = [
-            layers.LSTMCell(hidden_size) for i in range(num_layers)
-        ]
-
-    def attention(self, hidden, encoder_output, encoder_padding_mask):
-        query = layers.fc(
-            hidden, size=encoder_output.shape[-1], bias_attr=False
-        )
-        attn_scores = paddle.matmul(
-            layers.unsqueeze(query, [1]), encoder_output, transpose_y=True
-        )
-        if encoder_padding_mask is not None:
-            attn_scores = paddle.add(attn_scores, encoder_padding_mask)
-        attn_scores = paddle.nn.functional.softmax(attn_scores)
-        attn_out = paddle.squeeze(
-            paddle.matmul(attn_scores, encoder_output), [1]
-        )
-        attn_out = layers.concat([attn_out, hidden], 1)
-        attn_out = layers.fc(attn_out, size=self.hidden_size, bias_attr=False)
-        return attn_out
-
-    def call(
-        self, step_input, states, encoder_output, encoder_padding_mask=None
-    ):
-        lstm_states, input_feed = states
-        new_lstm_states = []
-        step_input = layers.concat([step_input, input_feed], 1)
-        for i in range(self.num_layers):
-            out, new_lstm_state = self.lstm_cells[i](step_input, lstm_states[i])
-            step_input = (
-                layers.dropout(out, self.dropout_prob)
-                if self.dropout_prob > 0
-                else out
-            )
-            new_lstm_states.append(new_lstm_state)
-        out = self.attention(step_input, encoder_output, encoder_padding_mask)
-        return out, [new_lstm_states, out]
-
-
-class Encoder:
-    def __init__(self, num_layers, hidden_size, dropout_prob=0.0):
-        self.encoder_cell = EncoderCell(num_layers, hidden_size, dropout_prob)
-
-    def __call__(self, src_emb, src_sequence_length):
-        encoder_output, encoder_final_state = layers.rnn(
-            cell=self.encoder_cell,
-            inputs=src_emb,
-            sequence_length=src_sequence_length,
-            is_reverse=False,
-        )
-        return encoder_output, encoder_final_state
-
-
-class Decoder:
-    def __init__(
-        self,
-        num_layers,
-        hidden_size,
-        dropout_prob,
-        decoding_strategy="infer_sample",
-        max_decoding_length=20,
-    ):
-        self.decoder_cell = DecoderCell(num_layers, hidden_size, dropout_prob)
-        self.decoding_strategy = decoding_strategy
-        self.max_decoding_length = (
-            None
-            if (self.decoding_strategy == "train_greedy")
-            else max_decoding_length
-        )
-
-    def __call__(
-        self,
-        decoder_initial_states,
-        encoder_output,
-        encoder_padding_mask,
-        **kwargs
-    ):
-        output_layer = kwargs.pop("output_layer", None)
-
-        beam_size = kwargs.get("beam_size", 4)
-        encoder_output = BeamSearchDecoder.tile_beam_merge_with_batch(
-            encoder_output, beam_size
-        )
-        encoder_padding_mask = BeamSearchDecoder.tile_beam_merge_with_batch(
-            encoder_padding_mask, beam_size
-        )
-        decoder = BeamSearchDecoder(
-            cell=self.decoder_cell, output_fn=output_layer, **kwargs
-        )
-
-        (
-            decoder_output,
-            decoder_final_state,
-            dec_seq_lengths,
-        ) = layers.dynamic_decode(
-            decoder,
-            inits=decoder_initial_states,
-            max_step_num=self.max_decoding_length,
-            encoder_output=encoder_output,
-            encoder_padding_mask=encoder_padding_mask,
-            impute_finished=False  # for test coverage
-            if self.decoding_strategy == "beam_search"
-            else True,
-            is_test=True if self.decoding_strategy == "beam_search" else False,
-            return_length=True,
-        )
-        return decoder_output, decoder_final_state, dec_seq_lengths
-
-
-class Seq2SeqModel:
-    """Seq2Seq model: RNN encoder-decoder with attention"""
-
-    def __init__(
-        self,
-        num_layers,
-        hidden_size,
-        dropout_prob,
-        src_vocab_size,
-        trg_vocab_size,
-        start_token,
-        end_token,
-        decoding_strategy="infer_sample",
-        max_decoding_length=20,
-        beam_size=4,
-    ):
-        self.start_token, self.end_token = start_token, end_token
-        self.max_decoding_length, self.beam_size = (
-            max_decoding_length,
-            beam_size,
-        )
-        self.src_embeder = paddle.nn.Embedding(
-            src_vocab_size,
-            hidden_size,
-            weight_attr=fluid.ParamAttr(name="source_embedding"),
-        )
-        self.trg_embeder = paddle.nn.Embedding(
-            trg_vocab_size,
-            hidden_size,
-            weight_attr=fluid.ParamAttr(name="target_embedding"),
-        )
-        self.encoder = Encoder(num_layers, hidden_size, dropout_prob)
-        self.decoder = Decoder(
-            num_layers,
-            hidden_size,
-            dropout_prob,
-            decoding_strategy,
-            max_decoding_length,
-        )
-        self.output_layer = lambda x: layers.fc(
-            x,
-            size=trg_vocab_size,
-            num_flatten_dims=len(x.shape) - 1,
-            param_attr=fluid.ParamAttr(),
-            bias_attr=False,
-        )
-
-    def __call__(self, src, src_length, trg=None, trg_length=None):
-        # encoder
-        encoder_output, encoder_final_state = self.encoder(
-            self.src_embeder(src), src_length
-        )
-
-        decoder_initial_states = [
-            encoder_final_state,
-            self.decoder.decoder_cell.get_initial_states(
-                batch_ref=encoder_output, shape=[encoder_output.shape[-1]]
-            ),
-        ]
-        src_mask = layers.sequence_mask(
-            src_length, maxlen=paddle.shape(src)[1], dtype="float32"
-        )
-        encoder_padding_mask = (src_mask - 1.0) * 1e9
-        encoder_padding_mask = layers.unsqueeze(encoder_padding_mask, [1])
-
-        # decoder
-        decoder_kwargs = (
-            {
-                "inputs": self.trg_embeder(trg),
-                "sequence_length": trg_length,
-            }
-            if self.decoder.decoding_strategy == "train_greedy"
-            else (
-                {
-                    "embedding_fn": self.trg_embeder,
-                    "beam_size": self.beam_size,
-                    "start_token": self.start_token,
-                    "end_token": self.end_token,
-                }
-                if self.decoder.decoding_strategy == "beam_search"
-                else {
-                    "embedding_fn": self.trg_embeder,
-                    "start_tokens": layers.fill_constant_batch_size_like(
-                        input=encoder_output,
-                        shape=[-1],
-                        dtype=src.dtype,
-                        value=self.start_token,
-                    ),
-                    "end_token": self.end_token,
-                }
-            )
-        )
-        decoder_kwargs["output_layer"] = self.output_layer
-
-        (decoder_output, decoder_final_state, dec_seq_lengths) = self.decoder(
-            decoder_initial_states,
-            encoder_output,
-            encoder_padding_mask,
-            **decoder_kwargs
-        )
-        if self.decoder.decoding_strategy == "beam_search":  # for inference
-            return decoder_output
-        logits, samples, sample_length = (
-            decoder_output.cell_outputs,
-            decoder_output.sample_ids,
-            dec_seq_lengths,
-        )
-        probs = paddle.nn.functional.softmax(logits)
-        return probs, samples, sample_length
 
 
 class PolicyGradient:
@@ -467,91 +214,6 @@ class SeqPGAgent:
             self.main_program, feed=feed_dict, fetch_list=fetch_list
         )
         return results
-
-
-class TestDynamicDecode(unittest.TestCase):
-    def setUp(self):
-        np.random.seed(123)
-        self.model_hparams = {
-            "num_layers": 2,
-            "hidden_size": 32,
-            "dropout_prob": 0.1,
-            "src_vocab_size": 100,
-            "trg_vocab_size": 100,
-            "start_token": 0,
-            "end_token": 1,
-            "decoding_strategy": "infer_greedy",
-            "max_decoding_length": 10,
-        }
-
-        self.iter_num = iter_num = 2
-        self.batch_size = batch_size = 4
-        src_seq_len = 10
-        trg_seq_len = 12
-        self.data = {
-            "src": np.random.randint(
-                2,
-                self.model_hparams["src_vocab_size"],
-                (iter_num * batch_size, src_seq_len),
-            ).astype("int64"),
-            "src_sequence_length": np.random.randint(
-                1, src_seq_len, (iter_num * batch_size,)
-            ).astype("int64"),
-            "trg": np.random.randint(
-                2,
-                self.model_hparams["src_vocab_size"],
-                (iter_num * batch_size, trg_seq_len),
-            ).astype("int64"),
-            "trg_sequence_length": np.random.randint(
-                1, trg_seq_len, (iter_num * batch_size,)
-            ).astype("int64"),
-            "label": np.random.randint(
-                2,
-                self.model_hparams["src_vocab_size"],
-                (iter_num * batch_size, trg_seq_len, 1),
-            ).astype("int64"),
-        }
-
-        place = (
-            core.CUDAPlace(0)
-            if core.is_compiled_with_cuda()
-            else core.CPUPlace()
-        )
-        self.exe = Executor(place)
-
-    def test_beam_search_infer(self):
-        paddle.set_default_dtype("float32")
-        paddle.enable_static()
-        self.model_hparams["decoding_strategy"] = "beam_search"
-        main_program = fluid.Program()
-        startup_program = fluid.Program()
-        with fluid.program_guard(main_program, startup_program):
-            source = fluid.data(name="src", shape=[None, None], dtype="int64")
-            source_length = fluid.data(
-                name="src_sequence_length", shape=[None], dtype="int64"
-            )
-            model = Seq2SeqModel(**self.model_hparams)
-            output = model(source, source_length)
-
-        self.exe.run(startup_program)
-        for iter_idx in range(self.iter_num):
-            trans_ids = self.exe.run(
-                program=main_program,
-                feed={
-                    "src": self.data["src"][
-                        iter_idx
-                        * self.batch_size : (iter_idx + 1)
-                        * self.batch_size,
-                        :,
-                    ],
-                    "src_sequence_length": self.data["src_sequence_length"][
-                        iter_idx
-                        * self.batch_size : (iter_idx + 1)
-                        * self.batch_size
-                    ],
-                },
-                fetch_list=[output],
-            )[0]
 
 
 class ModuleApiTest(unittest.TestCase):
